rename new causal vae
