+++ examples/opensora_pku/opensora/models/ae/videobase/causal_vae/modeling_causalvae.py
@@ -1,61 +1,144 @@
+import glob
+import json
 import logging
-
-import numpy as np
+import os
+from typing import Tuple
 
 import mindspore as ms
 from mindspore import nn, ops
 
-from ..modules.attention import make_attn
-from ..modules.conv import CausalConv3d
+from ..modeling_videobase import VideoBaseAE
+from ..modules.conv import CausalConv3d, Conv2d
 from ..modules.ops import nonlinearity
-from ..modules.resnet_block import ResnetBlock3D
-from ..modules.updownsample import SpatialDownsample2x, SpatialUpsample2x, TimeDownsample2x, TimeUpsample2x
+from ..utils.model_utils import resolve_str_to_obj
 
 logger = logging.getLogger(__name__)
 
 
-class CausalVAEModel(nn.Cell):
+class CausalVAEModel(VideoBaseAE):
+    """
+    The default vales are set to be the same as those used in OpenSora v1.1
+    """
+
     def __init__(
         self,
-        ddconfig,
-        embed_dim,
+        lr: float = 1e-5,  # ignore
+        hidden_size: int = 128,
+        z_channels: int = 4,
+        hidden_size_mult: Tuple[int] = (1, 2, 4, 4),
+        attn_resolutions: Tuple[int] = [],
+        dropout: float = 0.0,
+        resolution: int = 256,
+        double_z: bool = True,
+        embed_dim: int = 4,
+        num_res_blocks: int = 2,
+        loss_type: str = "opensora.models.ae.videobase.losses.LPIPSWithDiscriminator",  # ignore
+        loss_params: dict = {  # ignore
+            "kl_weight": 0.000001,
+            "logvar_init": 0.0,
+            "disc_start": 2001,
+            "disc_weight": 0.5,
+        },
+        q_conv: str = "CausalConv3d",
+        encoder_conv_in: str = "CausalConv3d",
+        encoder_conv_out: str = "CausalConv3d",
+        encoder_attention: str = "AttnBlock3D",
+        encoder_resnet_blocks: Tuple[str] = (
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+        ),
+        encoder_spatial_downsample: Tuple[str] = (
+            "SpatialDownsample2x",
+            "SpatialDownsample2x",
+            "SpatialDownsample2x",
+            "",
+        ),
+        encoder_temporal_downsample: Tuple[str] = (
+            "",
+            "TimeDownsample2x",
+            "TimeDownsample2x",
+            "",
+        ),
+        encoder_mid_resnet: str = "ResnetBlock3D",
+        decoder_conv_in: str = "CausalConv3d",
+        decoder_conv_out: str = "CausalConv3d",
+        decoder_attention: str = "AttnBlock3D",
+        decoder_resnet_blocks: Tuple[str] = (
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+        ),
+        decoder_spatial_upsample: Tuple[str] = (
+            "",
+            "SpatialUpsample2x",
+            "SpatialUpsample2x",
+            "SpatialUpsample2x",
+        ),
+        decoder_temporal_upsample: Tuple[str] = ("", "", "TimeUpsample2x", "TimeUpsample2x"),
+        decoder_mid_resnet: str = "ResnetBlock3D",
         ckpt_path=None,
         ignore_keys=[],
-        colorize_nlabels=None,
         monitor=None,
         use_fp16=False,
         upcast_sigmoid=False,
     ):
         super().__init__()
-        self.dtype = ms.float16 if use_fp16 else ms.float32
-        # print("D--: ddconfig: ", ddconfig)
+        dtype = ms.float16 if use_fp16 else ms.float32
 
-        self.encoder = Encoder(dtype=self.dtype, upcast_sigmoid=upcast_sigmoid, **ddconfig)
-        self.decoder = Decoder(dtype=self.dtype, upcast_sigmoid=upcast_sigmoid, **ddconfig)
-        assert ddconfig["double_z"]
-        if ddconfig["split_time_upsample"]:
-            logger.info("Exclude first frame from time upsample")
-        self.quant_conv = CausalConv3d(
-            2 * ddconfig["z_channels"],
-            2 * embed_dim,
-            1,
+        self.encoder = Encoder(
+            z_channels=z_channels,
+            hidden_size=hidden_size,
+            hidden_size_mult=hidden_size_mult,
+            attn_resolutions=attn_resolutions,
+            conv_in=encoder_conv_in,
+            conv_out=encoder_conv_out,
+            attention=encoder_attention,
+            resnet_blocks=encoder_resnet_blocks,
+            spatial_downsample=encoder_spatial_downsample,
+            temporal_downsample=encoder_temporal_downsample,
+            mid_resnet=encoder_mid_resnet,
+            dropout=dropout,
+            resolution=resolution,
+            num_res_blocks=num_res_blocks,
+            double_z=double_z,
+            dtype=dtype,
+            upcast_sigmoid=upcast_sigmoid,
         )
-        self.post_quant_conv = CausalConv3d(
-            embed_dim,
-            ddconfig["z_channels"],
-            1,
+
+        self.decoder = Decoder(
+            z_channels=z_channels,
+            hidden_size=hidden_size,
+            hidden_size_mult=hidden_size_mult,
+            attn_resolutions=attn_resolutions,
+            conv_in=decoder_conv_in,
+            conv_out=decoder_conv_out,
+            attention=decoder_attention,
+            resnet_blocks=decoder_resnet_blocks,
+            spatial_upsample=decoder_spatial_upsample,
+            temporal_upsample=decoder_temporal_upsample,
+            mid_resnet=decoder_mid_resnet,
+            dropout=dropout,
+            resolution=resolution,
+            num_res_blocks=num_res_blocks,
+            dtype=dtype,
+            upcast_sigmoid=upcast_sigmoid,
         )
+        quant_conv_cls = resolve_str_to_obj(q_conv)
+        self.quant_conv = quant_conv_cls(2 * z_channels, 2 * embed_dim, 1)
+        self.post_quant_conv = quant_conv_cls(embed_dim, z_channels, 1)
+
         self.embed_dim = embed_dim
 
-        if colorize_nlabels is not None:
-            assert type(colorize_nlabels) == int
-            self.register_buffer("colorize", ms.ops.standard_normal(3, colorize_nlabels, 1, 1))
         if monitor is not None:
             self.monitor = monitor
         if ckpt_path is not None:
             self.init_from_ckpt(ckpt_path, ignore_keys=ignore_keys)
 
         self.split = ops.Split(axis=1, output_num=2)
+        self.concat = ops.Concat(axis=1)
         self.exp = ops.Exp()
         self.stdnormal = ops.StandardNormal()
 
@@ -63,9 +146,9 @@ class CausalVAEModel(nn.Cell):
         # self.decoder.recompute()
         self.tile_sample_min_size = 256
         self.tile_sample_min_size_t = 65
-        hidden_size_mult = ddconfig["ch_mult"]
         self.tile_latent_min_size = int(self.tile_sample_min_size / (2 ** (len(hidden_size_mult) - 1)))
-        # self.tile_latent_min_size_t = int((self.tile_sample_min_size_t-1) / (2 ** self.time_compress)) + 1
+        t_down_ratio = [i for i in encoder_temporal_downsample if len(i) > 0]
+        self.tile_latent_min_size_t = int((self.tile_sample_min_size_t - 1) / (2 ** len(t_down_ratio))) + 1
         self.tile_overlap_factor = 0.25
         self.use_tiling = False
 
@@ -127,7 +210,7 @@ class CausalVAEModel(nn.Cell):
             if len(u) > 0:
                 logger.info("checkpoint param not loaded: ", u)
 
-    def init_from_ckpt(self, path, ignore_keys=list(), remove_prefix=["first_stage_model.", "autoencoder."]):
+    def init_from_ckpt(self, path, ignore_keys=list()):
         # TODO: support auto download pretrained checkpoints
         sd = ms.load_checkpoint(path)
         keys = list(sd.keys())
@@ -139,6 +222,32 @@ class CausalVAEModel(nn.Cell):
 
         ms.load_param_into_net(self, sd, strict_load=False)
         logger.info(f"Restored from {path}")
+
+    @classmethod  # rewrite class method to load
+    def from_pretrained(
+        cls, pretrained_model_path, subfolder=None, checkpoint_path=None, ignore_keys=["loss."], **kwargs
+    ):
+        if subfolder is not None:
+            pretrained_model_path = os.path.join(pretrained_model_path, subfolder)
+
+        config_file = os.path.join(pretrained_model_path, "config.json")
+        if not os.path.isfile(config_file):
+            raise RuntimeError(f"{config_file} does not exist")
+        with open(config_file, "r") as f:
+            config = json.load(f)
+
+        model = cls.from_config(config, **kwargs)
+        if checkpoint_path is None or len(checkpoint_path) == 0:
+            # search for ckpt under pretrained_model_path
+            ckpt_paths = glob.glob(os.path.join(pretrained_model_path, "*.ckpt"))
+            assert len(ckpt_paths) == 1, f"Expect to find one checkpoint file under {pretrained_model_path}"
+            f", but found {len(ckpt_paths)} files that end with `.ckpt`"
+            ckpt = ckpt_paths[0]
+        else:
+            ckpt = checkpoint_path
+        model.init_from_ckpt(ckpt, ignore_keys=ignore_keys)
+
+        return model
 
     def _encode(self, x):
         # return latent distribution, N(mean, logvar)
@@ -157,8 +266,12 @@ class CausalVAEModel(nn.Cell):
         return z
 
     def encode(self, x):
-        if self.use_tiling and (x.shape[-1] > self.tile_sample_min_size or x.shape[-2] > self.tile_sample_min_size):
-            posterior_mean, posterior_logvar = self.tiled_encode2d(x)
+        if self.use_tiling and (
+            x.shape[-1] > self.tile_sample_min_size
+            or x.shape[-2] > self.tile_sample_min_size
+            or x.shape[-3] > self.tile_sample_min_size_t
+        ):
+            posterior_mean, posterior_logvar = self.tiled_encode(x)
         else:
             # embedding, get latent representation z
             posterior_mean, posterior_logvar = self._encode(x)
@@ -204,9 +317,37 @@ class CausalVAEModel(nn.Cell):
         mean, logvar = self.split(moments)
         return mean, logvar
 
+    def tiled_encode(self, x):
+        t = x.shape[2]
+        t_chunk_idx = [i for i in range(0, t, self.tile_sample_min_size_t - 1)]
+        if len(t_chunk_idx) == 1 and t_chunk_idx[0] == 0:
+            t_chunk_start_end = [[0, t]]
+        else:
+            t_chunk_start_end = [[t_chunk_idx[i], t_chunk_idx[i + 1] + 1] for i in range(len(t_chunk_idx) - 1)]
+            if t_chunk_start_end[-1][-1] > t:
+                t_chunk_start_end[-1][-1] = t
+            elif t_chunk_start_end[-1][-1] < t:
+                last_start_end = [t_chunk_idx[-1], t]
+                t_chunk_start_end.append(last_start_end)
+        moments = []
+        for idx, (start, end) in enumerate(t_chunk_start_end):
+            chunk_x = x[:, :, start:end]
+            if idx != 0:
+                moment = self.concat(self.tiled_encode2d(chunk_x))[:, :, 1:]
+            else:
+                moment = self.concat(self.tiled_encode2d(chunk_x))
+            moments.append(moment)
+        moments = ops.cat(moments, axis=2)
+        mean, logvar = self.split(moments)
+        return mean, logvar
+
     def decode(self, z):
-        if self.use_tiling and (z.shape[-1] > self.tile_latent_min_size or z.shape[-2] > self.tile_latent_min_size):
-            return self.tiled_decode2d(z)
+        if self.use_tiling and (
+            z.shape[-1] > self.tile_latent_min_size
+            or z.shape[-2] > self.tile_latent_min_size
+            or z.shape[-3] > self.tile_latent_min_size_t
+        ):
+            return self.tiled_decode(z)
         z = self.post_quant_conv(z)
         dec = self.decoder(z)
         return dec
@@ -249,6 +390,29 @@ class CausalVAEModel(nn.Cell):
         dec = ops.cat(result_rows, axis=3)
         return dec
 
+    def tiled_decode(self, x):
+        t = x.shape[2]
+        t_chunk_idx = [i for i in range(0, t, self.tile_latent_min_size_t - 1)]
+        if len(t_chunk_idx) == 1 and t_chunk_idx[0] == 0:
+            t_chunk_start_end = [[0, t]]
+        else:
+            t_chunk_start_end = [[t_chunk_idx[i], t_chunk_idx[i + 1] + 1] for i in range(len(t_chunk_idx) - 1)]
+            if t_chunk_start_end[-1][-1] > t:
+                t_chunk_start_end[-1][-1] = t
+            elif t_chunk_start_end[-1][-1] < t:
+                last_start_end = [t_chunk_idx[-1], t]
+                t_chunk_start_end.append(last_start_end)
+        dec_ = []
+        for idx, (start, end) in enumerate(t_chunk_start_end):
+            chunk_x = x[:, :, start:end]
+            if idx != 0:
+                dec = self.tiled_decode2d(chunk_x)[:, :, 1:]
+            else:
+                dec = self.tiled_decode2d(chunk_x)
+            dec_.append(dec)
+        dec_ = ops.cat(dec_, axis=2)
+        return dec_
+
     def construct(self, input):
         # overall pass, mostly for training
         posterior_mean, posterior_logvar = self._encode(input)
@@ -280,73 +444,99 @@ class CausalVAEModel(nn.Cell):
             )
         return b
 
+    def validation_step(self, batch_idx):
+        raise NotImplementedError
+
 
 class Encoder(nn.Cell):
+    """
+    default value aligned to v1.1 vae config.json
+    """
+
     def __init__(
         self,
-        *,
-        ch,
-        out_ch,
-        ch_mult=(1, 2, 4, 4),
-        num_res_blocks,
-        attn_resolutions,
-        dropout=0.0,
-        resamp_with_conv=True,
-        in_channels,
-        resolution,
-        z_channels,
-        double_z=True,
-        use_linear_attn=False,
-        attn_type="vanilla3D",  # diff 3d
-        dtype=ms.float32,
-        time_compress=2,  # diff 3d
+        z_channels: int = 4,
+        hidden_size: int = 128,
+        hidden_size_mult: Tuple[int] = (1, 2, 4, 4),
+        attn_resolutions: Tuple[int] = (),
+        conv_in: str = "Conv2d",
+        conv_out: str = "CausalConv3d",
+        attention: str = "AttnBlock3D",  # already fixed, same as AttnBlock3DFix
+        resnet_blocks: Tuple[str] = (
+            "ResnetBlock2D",
+            "ResnetBlock2D",
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+        ),
+        spatial_downsample: Tuple[str] = (
+            "Downsample",
+            "Downsample",
+            "Downsample",
+            "",
+        ),
+        temporal_downsample: Tuple[str] = (
+            "",
+            "TimeDownsampleRes2x",
+            "TimeDownsampleRes2x",
+            "",
+        ),
+        mid_resnet: str = "ResnetBlock3D",
+        dropout: float = 0.0,
+        resolution: int = 256,
+        num_res_blocks: int = 2,
+        double_z: bool = True,
         upcast_sigmoid=False,
+        dtype=ms.float32,
         **ignore_kwargs,
     ):
         """
         ch: hidden size, i.e. output channels of the first conv layer. typical: 128
         out_ch: placeholder, not used in Encoder
-        ch_mult: channel multiply factors for each res block, also determine the number of res blocks.
-            Each block will be applied with spatial downsample x2 except for the last block. In total, the spatial downsample rate = 2**(len(ch_mult)-1)
+        hidden_size_mult: channel multiply factors for each res block, also determine the number of res blocks.
+            Each block will be applied with spatial downsample x2 except for the last block.
+            In total, the spatial downsample rate = 2**(len(hidden_size_mult)-1)
         resolution: spatial resolution, 256
-        time_compress: the begging `time_compress` blocks will be applied with temporal downsample x2. In total, the temporal downsample rate = 2**time_compress
+        time_compress: the begging `time_compress` blocks will be applied with temporal downsample x2.
+            In total, the temporal downsample rate = 2**time_compress
         """
-        # TODO: For input to AttnBlock3D, T=1, it's better to squeeze and use 2D AttnBlock
         super().__init__()
-        if use_linear_attn:
-            attn_type = "linear"
-        self.ch = ch
-        self.time_compress = time_compress
-        self.num_resolutions = len(ch_mult)
-        self.num_res_blocks = num_res_blocks
+        assert len(resnet_blocks) == len(hidden_size_mult), print(hidden_size_mult, resnet_blocks)
+        self.num_resolutions = len(hidden_size_mult)
         self.resolution = resolution
-        self.in_channels = in_channels
+        self.num_res_blocks = num_res_blocks
 
         self.dtype = dtype
         self.upcast_sigmoid = (upcast_sigmoid,)
 
-        # downsampling
-        # diff 3d
-        self.conv_in = CausalConv3d(
-            in_channels,
-            self.ch,
-            kernel_size=3,
-            stride=1,
-            padding=1,
-        )
+        # 1. Input conv
+        if conv_in == "Conv2d":
+            self.conv_in = Conv2d(3, hidden_size, kernel_size=3, stride=1, pad_mode="pad", padding=1, has_bias=True)
+        elif conv_in == "CausalConv3d":
+            self.conv_in = CausalConv3d(
+                3,
+                self.ch,
+                kernel_size=3,
+                stride=1,
+                padding=1,
+            )
+        else:
+            raise NotImplementedError
 
+        # 2. Downsample
         curr_res = resolution
-        in_ch_mult = (1,) + tuple(ch_mult)
+        in_ch_mult = (1,) + tuple(hidden_size_mult)
         self.in_ch_mult = in_ch_mult
         self.down = nn.CellList(auto_prefix=False)
+        self.downsample_flag = [0] * self.num_resolutions
+        self.time_downsample_flag = [0] * self.num_resolutions
         for i_level in range(self.num_resolutions):
             block = nn.CellList()
             attn = nn.CellList()
-            block_in = ch * in_ch_mult[i_level]  # input channels
-            block_out = ch * ch_mult[i_level]  # output channels
+            block_in = hidden_size * in_ch_mult[i_level]  # input channels
+            block_out = hidden_size * hidden_size_mult[i_level]  # output channels
             for i_block in range(self.num_res_blocks):
                 block.append(
-                    ResnetBlock3D(
+                    resolve_str_to_obj(resnet_blocks[i_level])(
                         in_channels=block_in,
                         out_channels=block_out,
                         dropout=dropout,
@@ -356,20 +546,28 @@ class Encoder(nn.Cell):
                 )
                 block_in = block_out
                 if curr_res in attn_resolutions:
-                    attn.append(make_attn(block_in, attn_type=attn_type, dtype=self.dtype))
+                    attn.append(resolve_str_to_obj(attention)(block_in, dtype=self.dtype))
+
             down = nn.Cell()
             down.block = block
             down.attn = attn
-            # do spatial downsample except for the last block
-            if i_level != self.num_resolutions - 1:
-                down.downsample = SpatialDownsample2x(block_in, block_in, dtype=self.dtype)
+
+            # do spatial downsample according to config
+            if spatial_downsample[i_level]:
+                down.downsample = resolve_str_to_obj(spatial_downsample[i_level])(block_in, block_in, dtype=self.dtype)
                 curr_res = curr_res // 2
+                self.downsample_flag[i_level] = 1
             else:
+                # TODO: still need it for 910b in new MS version?
                 down.downsample = nn.Identity()
-            # do temporal downsample for beginning tc blocks
-            if i_level < self.time_compress:
-                down.time_downsample = TimeDownsample2x()
+
+            # do temporal downsample according to config
+            if temporal_downsample[i_level]:
+                # TODO: add dtype support?
+                down.time_downsample = resolve_str_to_obj(temporal_downsample[i_level])(block_in, block_in)
+                self.time_downsample_flag[i_level] = 1
             else:
+                # TODO: still need it for 910b in new MS version?
                 down.time_downsample = nn.Identity()
 
             down.update_parameters_name(prefix=self.param_prefix + f"down.{i_level}.")
@@ -377,15 +575,15 @@ class Encoder(nn.Cell):
 
         # middle
         self.mid = nn.Cell()
-        self.mid.block_1 = ResnetBlock3D(
+        self.mid.block_1 = resolve_str_to_obj(mid_resnet)(
             in_channels=block_in,
             out_channels=block_in,
             dropout=dropout,
             dtype=self.dtype,
             upcast_sigmoid=upcast_sigmoid,
         )
-        self.mid.attn_1 = make_attn(block_in, attn_type=attn_type, dtype=self.dtype)
-        self.mid.block_2 = ResnetBlock3D(
+        self.mid.attn_1 = resolve_str_to_obj(attention)(block_in, dtype=self.dtype)
+        self.mid.block_2 = resolve_str_to_obj(mid_resnet)(
             in_channels=block_in,
             out_channels=block_in,
             dropout=dropout,
@@ -396,8 +594,10 @@ class Encoder(nn.Cell):
 
         # end
         self.norm_out = nn.GroupNorm(num_groups=32, num_channels=block_in, eps=1e-6, affine=True)
+        # self.norm_out = Normalize(block_in, extend=True)
 
-        self.conv_out = CausalConv3d(
+        assert conv_out == "CausalConv3d", "Only CausalConv3d is supported for conv_out"
+        self.conv_out = resolve_str_to_obj(conv_out)(
             block_in,
             2 * z_channels if double_z else z_channels,
             kernel_size=3,
@@ -410,17 +610,21 @@ class Encoder(nn.Cell):
         hs = [self.conv_in(x)]
         for i_level in range(self.num_resolutions):
             for i_block in range(self.num_res_blocks):
+                # import pdb; pdb.set_trace()
                 h = self.down[i_level].block[i_block](hs[-1])
                 if len(self.down[i_level].attn) > 0:
                     h = self.down[i_level].attn[i_block](h)
                 hs.append(h)
-            if i_level != self.num_resolutions - 1:
+            # if hasattr(self.down[i_level], "downsample"):
+            #    if not isinstance(self.down[i_level].downsample, nn.Identity):
+            if self.downsample_flag[i_level]:
                 hs.append(self.down[i_level].downsample(hs[-1]))
-            if i_level < self.time_compress:
-                hs.append(self.down[i_level].time_downsample(hs[-1]))
+            # if hasattr(self.down[i_level], "time_downsample"):
+            #    if not isinstance(self.down[i_level].time_downsample, nn.Identity):
+            if self.time_downsample_flag[i_level]:
+                hs_down = self.down[i_level].time_downsample(hs[-1])
+                hs.append(hs_down)
 
-        # import pdb
-        # pdb.set_trace()
         # middle
         h = hs[-1]
         h = self.mid.block_1(h)
@@ -435,70 +639,79 @@ class Encoder(nn.Cell):
 
 
 class Decoder(nn.Cell):
+    """
+    default value aligned to v1.1 vae config.json
+    """
+
     def __init__(
         self,
-        *,
-        ch,
-        out_ch,
-        ch_mult=(1, 2, 4, 8),
-        num_res_blocks,
-        attn_resolutions,
-        dropout=0.0,
-        resamp_with_conv=True,
-        in_channels,
-        resolution,
-        z_channels,
-        give_pre_end=False,
-        tanh_out=False,
-        use_linear_attn=False,
-        attn_type="vanilla3D",
-        time_compress=2,
-        split_time_upsample=True,  # TODO: ablate
-        dtype=ms.float32,
+        z_channels: int = 4,
+        hidden_size: int = 128,
+        hidden_size_mult: Tuple[int] = (1, 2, 4, 4),
+        attn_resolutions: Tuple[int] = (),
+        conv_in: str = "CausalConv3d",
+        conv_out: str = "CausalConv3d",
+        attention: str = "AttnBlock3D",  # already fixed, same as AttnBlock3DFix
+        resnet_blocks: Tuple[str] = (
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+            "ResnetBlock3D",
+        ),
+        spatial_upsample: Tuple[str] = ("", "SpatialUpsample2x", "SpatialUpsample2x", "SpatialUpsample2x"),
+        temporal_upsample: Tuple[str] = ("", "", "TimeUpsampleRes2x", "TimeUpsampleRes2x"),
+        mid_resnet: str = "ResnetBlock3D",
+        dropout: float = 0.0,
+        resolution: int = 256,
+        num_res_blocks: int = 2,
+        double_z: bool = True,
         upcast_sigmoid=False,
-        **ignorekwargs,
+        dtype=ms.float32,
+        **ignore_kwargs,
     ):
         super().__init__()
-        # if use_linear_attn: attn_type = "linear"
-        self.ch = ch
-        self.time_compress = time_compress
-        self.temb_ch = 0
-        self.num_resolutions = len(ch_mult)
-        self.num_res_blocks = num_res_blocks
+
+        self.num_resolutions = len(hidden_size_mult)
         self.resolution = resolution
-        self.in_channels = in_channels
-        self.give_pre_end = give_pre_end
-        self.tanh_out = tanh_out
+        self.num_res_blocks = num_res_blocks
+
         self.dtype = dtype
         self.upcast_sigmoid = upcast_sigmoid
 
+        # 1. decode input z conv
         # compute in_ch_mult, block_in and curr_res at lowest res
-        # in_ch_mult = (1,) + tuple(ch_mult)
-        block_in = ch * ch_mult[self.num_resolutions - 1]
+        block_in = hidden_size * hidden_size_mult[self.num_resolutions - 1]
         curr_res = resolution // 2 ** (self.num_resolutions - 1)
-        self.z_shape = (1, z_channels, curr_res, curr_res)
-        logger.info("Working with z of shape {} = {} dimensions.".format(self.z_shape, np.prod(self.z_shape)))
+        # self.z_shape = (1, z_channels, curr_res, curr_res)
+        # logger.info("Working with z of shape {} = {} dimensions.".format(self.z_shape, np.prod(self.z_shape)))
 
         # z to block_in
+        assert conv_in == "CausalConv3d", "Only CausalConv3d is supported for conv_in in Decoder currently"
         self.conv_in = CausalConv3d(z_channels, block_in, kernel_size=3, padding=1)
 
-        # middle
+        # 2. middle
         self.mid = nn.Cell()
-        self.mid.block_1 = ResnetBlock3D(in_channels=block_in, out_channels=block_in, dropout=dropout, dtype=self.dtype)
-        self.mid.attn_1 = make_attn(block_in, attn_type=attn_type, dtype=self.dtype)
-        self.mid.block_2 = ResnetBlock3D(in_channels=block_in, out_channels=block_in, dropout=dropout, dtype=self.dtype)
+        self.mid.block_1 = resolve_str_to_obj(mid_resnet)(
+            in_channels=block_in, out_channels=block_in, dropout=dropout, dtype=self.dtype
+        )
+        self.mid.attn_1 = resolve_str_to_obj(attention)(block_in, dtype=self.dtype)
+        self.mid.block_2 = resolve_str_to_obj(mid_resnet)(
+            in_channels=block_in, out_channels=block_in, dropout=dropout, dtype=self.dtype
+        )
         self.mid.update_parameters_name(prefix=self.param_prefix + "mid.")
 
-        # upsampling
+        # 3. upsampling
         self.up = nn.CellList(auto_prefix=False)
+        self.upsample_flag = [0] * self.num_resolutions
+        self.time_upsample_flag = [0] * self.num_resolutions
         # i_level: 3 -> 2 -> 1 -> 0
         for i_level in reversed(range(self.num_resolutions)):
             block = nn.CellList()
             attn = nn.CellList()
-            block_out = ch * ch_mult[i_level]
+            block_out = hidden_size * hidden_size_mult[i_level]
             for i_block in range(self.num_res_blocks + 1):
                 block.append(
-                    ResnetBlock3D(
+                    resolve_str_to_obj(resnet_blocks[i_level])(
                         in_channels=block_in,
                         out_channels=block_out,
                         dropout=dropout,
@@ -507,19 +720,22 @@ class Decoder(nn.Cell):
                 )
                 block_in = block_out
                 if curr_res in attn_resolutions:
-                    attn.append(make_attn(block_in, attn_type=attn_type, dtype=self.dtype))
+                    attn.append(resolve_str_to_obj(attention)(block_in, dtype=self.dtype))
             up = nn.Cell()
             up.block = block
             up.attn = attn
             # do spatial upsample x2 except for the first block
-            if i_level != 0:
-                up.upsample = SpatialUpsample2x(block_in, block_in, dtype=self.dtype)
+            if spatial_upsample[i_level]:
+                up.upsample = resolve_str_to_obj(spatial_upsample[i_level])(block_in, block_in, dtype=self.dtype)
                 curr_res = curr_res * 2
+                self.upsample_flag[i_level] = 1
             else:
                 up.upsample = nn.Identity()
-            # do temporal upsample x2 in the bottom tc blocks. TODO: choices for block positions to be injected with temporal upsample.
-            if i_level > self.num_resolutions - 1 - self.time_compress and i_level != 0:
-                up.time_upsample = TimeUpsample2x(exclude_first_frame=split_time_upsample)
+            # do temporal upsample x2 in the bottom tc blocks
+            if temporal_upsample[i_level]:
+                # TODO: support dtype?
+                up.time_upsample = resolve_str_to_obj(temporal_upsample[i_level])(block_in, block_in)
+                self.time_upsample_flag[i_level] = 1
             else:
                 up.time_upsample = nn.Identity()
 
@@ -531,8 +747,10 @@ class Decoder(nn.Cell):
 
         # end
         self.norm_out = nn.GroupNorm(num_groups=32, num_channels=block_in, eps=1e-6, affine=True)
+        # self.norm_out = Normalize(block_in, extend=True)
 
-        self.conv_out = CausalConv3d(block_in, out_ch, kernel_size=3, padding=1)
+        assert conv_out == "CausalConv3d", "Only CausalConv3d is supported for conv_out in Decoder currently"
+        self.conv_out = CausalConv3d(block_in, 3, kernel_size=3, padding=1)
 
     def construct(self, z):
         # z to block_in
@@ -551,18 +769,18 @@ class Decoder(nn.Cell):
                 h = self.up[i_level].block[i_block](h)
                 if len(self.up[i_level].attn) > 0:
                     h = self.up[i_level].attn[i_block](h)
-            if i_level != 0:
+            # if hasattr(self.up[i_level], 'upsample'):
+            #    if not isinstance(self.up[i_level].upsample, nn.Identity):
+            if self.upsample_flag[i_level]:
                 h = self.up[i_level].upsample(h)
 
-            if i_level > self.num_resolutions - 1 - self.time_compress and i_level != 0:
+            # if hasattr(self.up[i_level], 'time_upsample'):
+            #    if not isinstance(self.up[i_level].time_upsample, nn.Identity):
+            if self.time_upsample_flag[i_level]:
                 h = self.up[i_level].time_upsample(h)
 
         # end
-        if self.give_pre_end:
-            return h
         h = self.norm_out(h)
         h = nonlinearity(h, upcast=self.upcast_sigmoid)
         h = self.conv_out(h)
-        if self.tanh_out:
-            h = ops.tanh(h)
         return h
